NF: Draft of read support for NIfTI1 header extensions
This is all still rough and does nothing but storing raw extension
content. My main intention here is to get comments whether this is the
right way to do it (in terms of putting it into Image.extra and so on).

If that is the case, I'll move and update the corresponding unit tests
from the old codebase.