BF: Tentative fix for missing byteswapping of NIfTI1 extensions.
Unittests now pass on powerpc as well. Issue was that the previousNIfTI1
extension handling code completely ignored header endianness. Now it
relies on the endianness of the extensions being the same as for the
header (probably reasonable to assume).

There is still no endian-handling of actual extension content. I believe
that this should be handled by subclasses implementing particular
extension types, if necessary.