RF+TST - use apply scaling upcast for image load
Use safer apply_read_scaling routine to - er - apply the scaling from
the header to the read data.

Make tests to check for upcasting.  The tests revealed that the
upcasting was mostly happening already due to accidental float casting
when getting the slope and intercept for nifti1 and spm2analyze.

Expand tests of the underly int_scinter_ftype routine