RF - use full range for scaling
I was previously using the range calculated from the shared range. This
was to prevent the values overflowing into inf, for float -> int save,
followed by int -> float load (overflow on applying the scaling on
load). However, the apply_read_scaling deals with this problem, so the
full range is better (because it doesn't truncate the range) and is also
easier to understand.