RF - reverse arg order for arraywriter factory
The order was has_intercept, has_slope, because the default for both was
true, and it is more common to not want an intercept, than not want a
slope.

However, (slope, inter) are elsewhere in that order (for example in the
class names and in ``get_slope_inter``). I changed the argument order to
fit that convention.