ENH + TEST: enforce float image affines
Image affines could be any data type (depending on what was passed).
This could be odd when doing things like ``img.get_affine()[0, 0] = 1.1;
the value would be truncated if the original affine was an integer type.
Obviously this is numpy's default behavior, but we're copying the affine
anyway, so casting to float seems ordinary and a bit less surprising.