RF: check fp error in offset for offset scaling
When trying to scale integers only with inter, we need to make sure the
inter as cast to the scaling type does in fact keep the integers in the
range of the output type.  We also want to make sure that inter
generates numbers with the smallest possible absolute magnitude to
avoid floating point error in the resulting numbers.