BF: try other checks for badly-detected longdouble
Numpy gives wrong values from the np.finfo routine for PPC, and I put a
hack in to detect that.  It looks like numpy also gives wrong values on
Debian on the IBM POWER7, and that might have a real binary128
longdouble. Try to make the checks more flexible to recognize both
cases.