Update nibabel/analyze.py
Updated the way user provided slope and inter are handled when writing to file.
Taking into account the algebraic use of those quantities it is possible to integrate the previous use of to_file_map (i.e. conversion of float to int images) with user provided slope,inter.
Added a brief sanity check that raises an exception when numerical instability is caused by such values (The exception message could be more descriptive, but should do as a proof of concept).