RF: remove MGH get_affine method as duplicate
The behavior of the other image classes is to store the affine in the
image, and return than from ``get_affine``.  The affine can be modified
too.  The ``update_header`` method makes sure the affine for the image
matches the affine for the header.  Change MGH to match behavior of the
other image types (so falling back to SpatialImage.get_affine)