RF: simplify and generalize run_mod_cmd
`run_mod_cmd` was getting complicated and it was failing on Windows.
Because it was complicated I couldn't work out why it was failing.
Simplify by writing the script in the temporary directory rather than
trying to use a long command string.  The refactoring may mean that
paths with spaces will work.