RF: enforce mode='r' for opening HDF5 files
For some reason, the tests were running fine with the default assumption
that h5py.File was ``mode='r'`` by default, when run on their own
(``nosetests nibabel/tests/test_proxy_api.py``) but not when run as part
of the full test suite (``nosetests nibabel``).  Maybe there's a global
flag in the HDF library somewhere that was getting flipped.