API: set offset to zero when creating images
For Analyze types, treat the file offset as consumable. So, set to zero
when creating an image object, and set to the necessary (mimumum) offset
when writing the header.

If the user sets the offset directly to a value other than zero, use
that value or raise and error if it is too small.