BF: fix missed extension status check on Python 3
Python 3 was doing the check wrong for extension status == 0; integer
slicing into a byte array results in an integer, so comparison of this
integer against a byte string was always false.  This is OK as long as
the offset is default, but gives an error later if the offset is higher
than the default minimum.