API+BRK+TST: change ArrayWriter nan2zero API
We need to know nan2zero when calculating the scaling for arrays, so we
need nan2zero at the array writer __init__, and we can't use it when
calling `to_fileobj` as we have been up until now, because it would
involve recalculating the scalefactor on the fly, which would break the
API too.

I'm guessing no-one is using nan2zero=False as an argument to
`to_fileobj`, but in case they are, check if it's the same as the
__init__ value, raise an error if not, deprecation warning if so.

The tests are still broken because I haven't fixed the scaling yet.