API+TST: errors for bad scaling in array_to_file
It's time consuming to test for NaNs when converting from integers to
integers, and we don't want to check whether nan2zero value is within
range if there can be no NaNs, so raise errors for scaling that will
generate NaNs.  Next we add tests to confirm that we're not checking the
nan2zero nan_fill value for int to int scaling.