RF: move image scaling tests to mixin class
Move the scaling tests in the Analyze test base class into the Spm99
scaling mixin class, to make refactoring easier and the logic more
clear.  There were no tests that could be run for Analyze images, so it
was confusing to have the test method in the Analyze test class.