TST: tests for read_img_data function
Test the behavior of the `read_img_data` function as of commit 6d8f6c6.
This is before the changes to Analyze image types to make the header
slope, intercept and data offset be consumed during write, and reset to
default values.  These tests pass against that commit but fail against
current master, which is 62f4c70.