RF: revert fix allowing upper case file extensions
I thought about this for a while, but I don't like it, because it
changes the behavior of nibabel globally. Previously if you asked to
open ``my_image.NII`` it would barf, now it might find an image when it
previously it did not.  This is probably what the user intended, but
it's still a guess.  It's not hard to specify the upper case file
extension on systems that need it.

The next commit adds an explicit function to ignore extension case for
parrec2nii.