RF+TST: refactor / rename / test truncation checks
The function that was `_calc_extras` is now effectively just doing
truncation checks; rename accordingly.

The truncation tests were being used to knock volumes off the end of the
data shape, but I'm using the ``vol_is_full`` function for this now.
Refactor the checks to reflect fact that the check is not doing this
calculation.

Add checks for the number of b values and gradient directions.

Remove check for number of diffusion slices for now; sorry Eric - I
don't understand the check, and the comments worried me.  Can we wait
for test data to put that back?

In general I suppose it's possible to have any nomber of DTI volumes
with a given b value and gradient id number, so I don't at the moment
see how we can easily use these to check the predicted number of
volumes, but it's quite possible I didn't understand the check properly.

I have no idea why, but I found it very difficult to grasp what the
``must_exceed_one`` function parameter was for.  I think it has the
effect, if True, of disabling the expected vs actual number check if the
number of expected values is 0 or 1.

Because I found it confusing and because all the files I had did have
one value when the expected number of values was one, I didn't port that
check - maybe add it back if we find a counter-example?