NF: add context manager to clear warnings registry
Testing warnings can be really annoying, because once the warning has
been raises, by default, it cannot be raised again, because it is
already in the warnings registry.

Here is a context manager to reset the warnings registry inside the
context manager.

The general idea was inspired by the conversation here:

https://stackoverflow.com/questions/2390766/how-do-i-disable-and-then-re-enable-a-warning