RF+BF: '' from `get_nibabel_data()` when missing
Return empty string from ``get_nibabel_data()`` if directory does not
exist, instead of None.

Doing this means we can use os.path.join on the output even if it the
directory doesn't exist. The resulting filenames don't make sense, but
the decorator checks for this in any case, so the filenames don't get
used.