BF+ENH: Improve elem container functionality.
Allow addition on ElemList objects. Make both ElemList and ElemDict
work better with other objects of the same type (e.g. allow the
contructor to take them, allow ElemDict.update() to take ElemDict,
etc).

No longer try to autoconvert nested dict objects to ElemDict objects
as this is a bad idea and we do want to allow plain dicts to be
nested.