NF: allow truncated last track in trackvis file
From report on mailing list by Carolyn D. Langen - trackvis reader was
giving TypeError when trying to read file where last track was shorter
than declared in the 'n_points' data.

Allow this situation for the last track with False argument to new
`strict` kwarg of the trackvis `read` function.

This is a slight API break because:

* We are now raising a DataError if there are too few streamlines in the
  file, instead of a HeaderError;
* We are raising a DataError if the track is truncated, rather than a
  TypeError when trying to create the points array.