RF: use new cached append for multi-seq read
Use the new cached append to replace the coroutines for the read across
multiple array sequences in ``create_arraysequences_from_generator``.

Before this change (using coroutines):

```
Old: Loaded 5,000 streamlines in   2.85

New: Loaded 5,000 streamlines in    3.6
Speedup of 0.79
Old: Loaded 5,000 streamlines with scalars in   5.16
New: Loaded 5,000 streamlines with scalars in   7.13
Speedup of 0.723703
```

After this change (using cached append):

```
Old: Loaded 5,000 streamlines in   3.21

New: Loaded 5,000 streamlines in    3.9
Speedup of 0.82
Old: Loaded 5,000 streamlines with scalars in   5.21
New: Loaded 5,000 streamlines with scalars in   7.16
Speedup of 0.727654
```

This seems to be well within run-to-run measurement error.