RF+TST: drop CIFTI Matrix check, add tests
CIFTI matrix check was incorrect, metadata never required.  However, we
do need to have one or more MatrixIndicesMap entries referring to all
the axes in the data matrix.  Need to think about how to do this check,
as the XML generator does not know how many data dimensions we have.