BF(workaround): skip test_diff if hypothesis is not importable
Apparently hypothesis  is quite a tricky beast.  It might either fail to install
on elderly systems from source tarball, or even if it is installable, apparently
might rely on some coverage API which is present only in more recent versions
and thus fail upon import.
All those are of relevance for older systems and only for testing.