RF: minc - delay import of h5py until needed.
On Debian systems h5py comes with MPI enabled build.
I guess on some misconfigured systems, like my laptop, it might take awhile to import.

And since h5py here is imported upon import of `nibabel`, it makes
it long(er) to import it.  Here is timing on my laptop before the change:

    $> time python -c 'import nibabel'
    python -c 'import nibabel'  0.20s user 0.08s system 5% cpu 5.356 total

and here is after the change:

    $> time python -c 'import nibabel'
    python -c 'import nibabel'  0.13s user 0.02s system 100% cpu 0.150 total