RF: Add generic NiftiExtension base class
Nifti1Extension is a non-ideal base class for NIfTI extensions because
it assumes that it is safe to store use a null transformation, and thus
default to `bytes` objects. This makes it difficult to define its typing
behavior in a way that allows subclasses to refine the type such that
type-checkers understand it.

This patch creates a generic `NiftiExtension` class that parameterizes
the "runtime representation" type. Nifti1Extension subclasses with
another parameter that defaults to `bytes`, allowing it to be subclassed
in turn (preserving the Nifti1Extension -> Nifti1DicomExtension subclass
relationship) while still emitting `bytes`.

We could have simply made `Nifti1Extension` the base class, but the
mangle/unmangle methods need some casts or ignore comments to type-check
cleanly. This separation allows us to have a clean base class with the
legacy hacks cordoned off into an subclass.