BF+TST: Test and fix a bunch of multiframe fixes
Corrects issue where order of slice indices was assumed to match
the order needed to move along the direction of the slice normal,
which resulted in slice orientation flips.

Ignores indices that don't evenly divide data, and at the end will
try to combine those indices (if needed) into a single tuple index.