ENH: Allow path-like objects for tractogram load/saving
Allow path-like objects for tractogram load/saving.

Crucially, this patch set uses the capabilities of `pathlib.Path` to get
the suffix of a filename when detecing the format instead of relying on
the `splitext` function of the `os.path` module, which requires a `str`
instance.

Add the corresponding tests.

Change the docstrings to mark the types of the filenames as `path-like`
objects instead of only `str` objects.